+++ src/lib/services/validator.ts
@@ -4,8 +4,17 @@ import { z } from "zod";
 
 import { CURR_THEME } from "env";
 import type { Option, StakingShare, Validator, ValidatorAddr } from "lib/types";
-import { BlockVote, zBig, zUtcDate, zValidatorData } from "lib/types";
+import {
+  BlockVote,
+  zBig,
+  zCoin,
+  zUtcDate,
+  zValidatorAddr,
+  zValidatorData,
+} from "lib/types";
 import { parseWithError, removeSpecialChars, snakeToCamel } from "lib/utils";
+
+import { zBlocksResponse } from "./block";
 
 interface ValidatorResponse {
   operator_address: ValidatorAddr;
@@ -83,21 +92,23 @@ const zHistoricalPowersItem = z.object({
   voting_power: z.number().nonnegative(),
 });
 
-export const zHistoricalPowers = z
+export const zHistoricalPowersResponse = z
   .object({
     items: z.array(zHistoricalPowersItem),
     total: z.number(),
   })
-  .transform((data) => snakeToCamel(data));
-export type HistoricalPowers = z.infer<typeof zHistoricalPowers>;
+  .transform(snakeToCamel);
+export type HistoricalPowersResponse = z.infer<
+  typeof zHistoricalPowersResponse
+>;
 
 export const getHistoricalPowers = async (
   endpoint: string,
   validatorAddr: ValidatorAddr
-): Promise<HistoricalPowers> =>
+): Promise<HistoricalPowersResponse> =>
   axios
     .get(`${endpoint}/${validatorAddr}/historical-powers`)
-    .then(({ data }) => parseWithError(zHistoricalPowers, data));
+    .then(({ data }) => parseWithError(zHistoricalPowersResponse, data));
 
 const zValidatorsResponse = z
   .object({
@@ -184,3 +195,60 @@ export const getValidatorUptime = async (
       },
     })
     .then(({ data }) => parseWithError(zValidatorUptimeResponse, data));
+
+const zValidatorDelegationRelatedTxsResponseItem = z
+  .object({
+    tx_hash: z.string(),
+    height: z.number().positive(),
+    tokens: zCoin.array(),
+    timestamp: zUtcDate,
+    validator_address: zValidatorAddr,
+  })
+  .transform(snakeToCamel);
+
+const zValidatorDelegationRelatedTxsResponse = z.object({
+  items: z.array(zValidatorDelegationRelatedTxsResponseItem),
+  total: z.number().nonnegative(),
+});
+
+export type ValidatorDelegationRelatedTxsResponse = z.infer<
+  typeof zValidatorDelegationRelatedTxsResponse
+>;
+
+export const getValidatorDelegationRelatedTxs = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr,
+  limit: number,
+  offset: number
+) =>
+  axios
+    .get(
+      `${endpoint}/${encodeURIComponent(validatorAddress)}/delegation-related-txs`,
+      {
+        params: {
+          limit,
+          offset,
+        },
+      }
+    )
+    .then(({ data }) =>
+      parseWithError(zValidatorDelegationRelatedTxsResponse, data)
+    );
+
+export const getValidatorProposedBlocks = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr,
+  limit: number,
+  offset: number
+) =>
+  axios
+    .get(
+      `${endpoint}/${encodeURIComponent(validatorAddress)}/proposed-blocks`,
+      {
+        params: {
+          limit,
+          offset,
+        },
+      }
+    )
+    .then(({ data }) => parseWithError(zBlocksResponse, data));